% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colorCircle}
\alias{colorCircle}
\title{Plot a chord diagram of color transitions across the clip}
\usage{
colorCircle(frames, extra = FALSE)
}
\arguments{
\item{frames}{list of frames}

\item{extra}{plot intermediate colors}
}
\value{
none
}
\description{
Plot a chord diagram of color transitions across the clip
}
\examples{
print("example")

}
