% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{extractFramePalette}
\alias{extractFramePalette}
\title{extractFramePalette}
\usage{
extractFramePalette(
  img_path = NULL,
  palette_dim = 10,
  max_res = 780,
  graphics = TRUE
)
}
\arguments{
\item{jpegFramePath}{jpegFramePath}

\item{paletteDim}{paletteDim}

\item{title}{title}

\item{subtitle}{subtitle}
}
\value{
none
}
\description{
extractFramePalette
}
\examples{
print("example")

}
