% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{fileName}
\alias{fileName}
\title{fileName}
\usage{
fileName(name, ext = "png")
}
\arguments{
\item{name}{name}

\item{ext}{ext}
}
\value{
name
}
\description{
fileName
}
\examples{
print("example")

}
