% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{getFrames}
\alias{getFrames}
\title{getFrames}
\usage{
getFrames(orderBySeason = FALSE)
}
\arguments{
\item{orderBySeason}{orderBySeason}
}
\value{
frames
}
\description{
getFrames
}
\examples{
print("example")

}
