% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{getHueRef}
\alias{getHueRef}
\title{GetHueRef}
\usage{
getHueRef(hexString, extra = FALSE)
}
\arguments{
\item{hexString}{hexString}

\item{extra}{plot intermediate colors}
}
\value{
color name
}
\description{
GetHueRef
}
\examples{
print("example")

}
