% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{getSummary}
\alias{getSummary}
\title{getSummary}
\usage{
getSummary(framesCollection)
}
\arguments{
\item{framesCollection}{frames}
}
\value{
summary
}
\description{
getSummary
}
\examples{
print("example")

}
