% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{groupframes}
\alias{groupframes}
\title{groupframes}
\usage{
groupframes(frames, seconds = NULL, fps = chromaRenv$fps)
}
\arguments{
\item{frames}{frames}

\item{seconds}{seconds}

\item{fps}{fps}
}
\value{
frames
}
\description{
groupframes
}
\examples{
print("example")

}
