% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hslGet}
\alias{hslGet}
\title{Return H/S/V value of a hex color}
\usage{
hslGet(hexString, channel = c("H", "S", "V"))
}
\arguments{
\item{hexString}{hex color string}

\item{channel}{channel to return}
}
\value{
H/S/V channel value
}
\description{
Return H/S/V value of a hex color
}
\examples{
print("example")

}
