% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hueOnly}
\alias{hueOnly}
\title{Color correction to focus on hue only}
\usage{
hueOnly(hexString)
}
\arguments{
\item{hexString}{hex color string}
}
\value{
corrected hex color
}
\description{
Color correction to focus on hue only
}
\examples{
print("example")

}
