% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{isWarm}
\alias{isWarm}
\title{Get warmness of a color}
\usage{
isWarm(hexString)
}
\arguments{
\item{hexString}{hex color string}
}
\value{
temperature WARM/COLD/NEUTRAL and distance in degrees
}
\description{
Get warmness of a color and its distance from pure red
}
\examples{
print("example")

}
