% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{luminance}
\alias{luminance}
\title{Extract the luminance value of an RGB color}
\usage{
luminance(r, g, b)
}
\arguments{
\item{r}{Red channel}

\item{g}{Green channel}

\item{b}{Blue channel}
}
\value{
Luminance value
}
\description{
Extract the luminance value of an RGB color
}
\examples{
print("example")

}
