% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{lumOnly}
\alias{lumOnly}
\title{Color correction to focus on brightness only}
\usage{
lumOnly(hexString)
}
\arguments{
\item{hexString}{hex color string}
}
\value{
corrected hex color
}
\description{
Color correction to focus on brightness only
}
\examples{
print("example")

}
