\name{matrix}
\alias{matrix}
\docType{data}
\title{
matrix
}
\description{
description
}
\usage{data("matrix")}
\format{
  The format is:
List of 3
 $ :'data.frame':	196160 obs. of  6 variables:
  ..$ R      : num [1:196160] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:196160] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:196160] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:196160] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:196160] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:196160] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#292A22"
  ..- attr(*, "title")= chr "s1e1"
  ..- attr(*, "duration")= num 8173
 $ :'data.frame':	198950 obs. of  6 variables:
  ..$ R      : num [1:198950] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:198950] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:198950] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:198950] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:198950] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:198950] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#2D312A"
  ..- attr(*, "title")= chr "s1e2"
  ..- attr(*, "duration")= num 8290
 $ :'data.frame':	185945 obs. of  6 variables:
  ..$ R      : num [1:185945] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:185945] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:185945] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:185945] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:185945] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:185945] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#252A29"
  ..- attr(*, "title")= chr "s1e3"
  ..- attr(*, "duration")= num 7748
}
\details{
details
}
\source{
source
}
\references{
references
}
\examples{
data(matrix)
}
\keyword{datasets}
