\name{miyazaki}
\alias{miyazaki}
\docType{data}
\title{
miyazaki
}
\description{
description
}
\usage{data("miyazaki")}
\format{
  The format is:
List of 10
 $ :'data.frame':	168361 obs. of  6 variables:
  ..$ R      : num [1:168361] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:168361] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:168361] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:168361] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:168361] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:168361] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#5F544C"
  ..- attr(*, "title")= chr "1984 - Nausicaa of the Valley of the Wind"
  ..- attr(*, "duration")= num 7015
 $ :'data.frame':	179186 obs. of  6 variables:
  ..$ R      : num [1:179186] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:179186] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:179186] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:179186] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:179186] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:179186] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#464846"
  ..- attr(*, "title")= chr "1986 - Laputa Castle in the Sky"
  ..- attr(*, "duration")= num 7466
 $ :'data.frame':	124321 obs. of  6 variables:
  ..$ R      : num [1:124321] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:124321] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:124321] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:124321] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:124321] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:124321] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#4B524A"
  ..- attr(*, "title")= chr "1988 - My Neighbor Totoro"
  ..- attr(*, "duration")= num 5180
 $ :'data.frame':	134264 obs. of  6 variables:
  ..$ R      : num [1:134264] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:134264] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:134264] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:134264] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:134264] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:134264] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#555D58"
  ..- attr(*, "title")= chr "1992 - Porco Rosso"
  ..- attr(*, "duration")= num 5594
 $ :'data.frame':	191878 obs. of  6 variables:
  ..$ R      : num [1:191878] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:191878] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:191878] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:191878] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:191878] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:191878] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#3A413D"
  ..- attr(*, "title")= chr "1997 - Princess Mononoke"
  ..- attr(*, "duration")= num 7995
 $ :'data.frame':	148236 obs. of  6 variables:
  ..$ R      : num [1:148236] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:148236] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:148236] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:148236] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:148236] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:148236] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#4D575B"
  ..- attr(*, "title")= chr "1998 - Kiki's Delivery Service"
  ..- attr(*, "duration")= num 6176
 $ :'data.frame':	179221 obs. of  6 variables:
  ..$ R      : num [1:179221] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:179221] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:179221] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:179221] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:179221] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:179221] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#5F5343"
  ..- attr(*, "title")= chr "2001  - Spirited Away"
  ..- attr(*, "duration")= num 7468
 $ :'data.frame':	171409 obs. of  6 variables:
  ..$ R      : num [1:171409] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:171409] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:171409] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:171409] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:171409] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:171409] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#514D43"
  ..- attr(*, "title")= chr "2005 - Howl's Moving Castle"
  ..- attr(*, "duration")= num 7142
 $ :'data.frame':	145055 obs. of  6 variables:
  ..$ R      : num [1:145055] 0 0 0 0 0 ...
  ..$ G      : num [1:145055] 0 0 0 0 0 ...
  ..$ B      : num [1:145055] 0 0 0 0 0 ...
  ..$ hex    : chr [1:145055] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:145055] 0 0 0 0 0 ...
  ..$ frameId: int [1:145055] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#697060"
  ..- attr(*, "title")= chr "2008 - Ponyo on the Cliff by the Sea"
  ..- attr(*, "duration")= num 6044
 $ :'data.frame':	181985 obs. of  6 variables:
  ..$ R      : num [1:181985] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ G      : num [1:181985] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ B      : num [1:181985] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ hex    : chr [1:181985] "#000000" "#000000" "#000000" "#000000" ...
  ..$ lum    : num [1:181985] 0 0 0 0 0 0 0 0 0 0 ...
  ..$ frameId: int [1:181985] 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "avgRGB")= chr "#5D6150"
  ..- attr(*, "title")= chr "2013 - The Wind Rises"
  ..- attr(*, "duration")= num 7583
}
\details{
details
}
\source{
source
}
\references{
references
}
\examples{
data(miyazaki)
}
\keyword{datasets}
