% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotChannel}
\alias{plotChannel}
\title{plotChannel}
\usage{
plotChannel(frames, channel = c("h", "s", "l", "r", "g", "b"),
  npoly = c(1, 2, 3))
}
\arguments{
\item{frames}{frames}

\item{channel}{hsl}

\item{npoly}{npoly}
}
\value{
none
}
\description{
plotChannel
}
\examples{
print("example")

}
