% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotFrameline}
\alias{plotFrameline}
\title{plotFrameline}
\usage{
plotFrameline(
  frames,
  verbose = 0,
  summary = TRUE,
  vivid = FALSE,
  timeScale = TRUE,
  title = "MOVIE FRAMELINE",
  subtitle = "frameline"
)
}
\arguments{
\item{frames}{frames}

\item{verbose}{frames}

\item{summary}{frames}

\item{vivid}{frames}

\item{timeScale}{frames}

\item{title}{frames}

\item{subtitle}{frames}
}
\value{
none
}
\description{
plotFrameline
}
\examples{
print("example")

}
