% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotMultiFrameline}
\alias{plotMultiFrameline}
\title{plotMultiFrameline}
\usage{
plotMultiFrameline(
  framescollection,
  verbose = 0,
  summary = TRUE,
  vivid = FALSE,
  timeScale = TRUE,
  title = "MOVIES FRAMELINES",
  subtitle = "frameline"
)
}
\arguments{
\item{framescollection}{frames}

\item{verbose}{frames}

\item{summary}{frames}

\item{vivid}{frames}

\item{timeScale}{frames}

\item{title}{frames}

\item{subtitle}{frames}
}
\value{
none
}
\description{
plotMultiFrameline
}
\examples{
print("example")

}
