% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotSingleFrameline}
\alias{plotSingleFrameline}
\title{plotSingleFrameline}
\usage{
plotSingleFrameline(frames, verbose = 0, summary = TRUE,
  vivid = FALSE, title = "MOVIE FRAMELINE", subtitle = "frameline",
  xlim = NULL)
}
\arguments{
\item{frames}{frames}

\item{verbose}{frames}

\item{summary}{frames}

\item{vivid}{frames}

\item{title}{frames}

\item{subtitle}{frames}

\item{xlim}{frames}
}
\value{
none
}
\description{
plotSingleFrameline
}
\examples{
print("example")

}
