% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotTilesSummary}
\alias{plotTilesSummary}
\title{plotMultiFrameline}
\usage{
plotTilesSummary(
  summary,
  mode = "",
  verbose = 0,
  title = "MOVIE SUMMARY",
  subtitle = ""
)
}
\arguments{
\item{summary}{summary}

\item{mode}{mode}

\item{verbose}{verbose}

\item{title}{frames}

\item{subtitle}{frames}
}
\value{
none
}
\description{
plotMultiFrameline
}
\examples{
print("example")

}
