% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotTimeWindows}
\alias{plotTimeWindows}
\title{plotTimeWindows}
\usage{
plotTimeWindows(
  frames = NULL,
  vivid = TRUE,
  verbose = 1,
  title = "Frameline Time Windows",
  subtitle = "",
  left = "window size [seconds]"
)
}
\arguments{
\item{frames}{frames}

\item{vivid}{vivid}

\item{verbose}{verbose}

\item{title}{title}

\item{subtitle}{subtitle}

\item{left}{left}
}
\value{
none
}
\description{
plotTimeWindows
}
\examples{
print("example")

}
