% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{satOnly}
\alias{satOnly}
\title{Color correction to focus on saturation only}
\usage{
satOnly(hexString)
}
\arguments{
\item{hexString}{hex color string}
}
\value{
corrected hex color
}
\description{
Color correction to focus on saturation only
}
\examples{
print("example")

}
