% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{theme_semiVoid}
\alias{theme_semiVoid}
\title{theme_semiVoid}
\usage{
theme_semiVoid(xaxis = FALSE)
}
\arguments{
\item{xaxis}{xaxis}
}
\value{
theme
}
\description{
theme_semiVoid
}
\examples{
print("example")

}
