% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{vividHex}
\alias{vividHex}
\title{Make colors more vivid}
\usage{
vividHex(hexString, intensity = c("light", "medium", "strong", "ultra"))
}
\arguments{
\item{hexString}{hex color string}

\item{intensity}{intensity of the color correction}
}
\value{
corrected hex color
}
\description{
Make hex colors more vivid increasing saturation and brightness
}
\examples{
print("example")

}
