% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{extractFramePalette}
\alias{extractFramePalette}
\title{extractFramePalette}
\usage{
extractFramePalette(jpegFramePath = NULL, paletteDim = 10,
  title = "FRAME ANALYSIS", subtitle = "Palette")
}
\arguments{
\item{jpegFramePath}{jpegFramePath}

\item{paletteDim}{paletteDim}

\item{title}{title}

\item{subtitle}{subtitle}
}
\value{
none
}
\description{
extractFramePalette
}
\examples{
print("example")

}
