% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{extractFramePalette}
\alias{extractFramePalette}
\title{extractFramePalette}
\usage{
extractFramePalette(
  img_path = NULL,
  palette_dim = 10,
  max_res = 780,
  graphics = TRUE
)
}
\arguments{
\item{img_path}{img_path}

\item{palette_dim}{palette_dim}

\item{max_res}{max_res}

\item{graphics}{graphics}
}
\value{
none
}
\description{
extractFramePalette
}
\examples{
print("example")

}
