% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{getFrames}
\alias{getFrames}
\title{getFrames}
\usage{
getFrames(paths = c(NULL), names = c(NULL), orderBySeason = FALSE)
}
\arguments{
\item{paths}{files full path. Optional in Windows}

\item{names}{custom name for imported clips. Optional.}

\item{orderBySeason}{orderBySeason. FALSE by default.}
}
\value{
frames
}
\description{
getFrames
}
\examples{
print("example")

}
