% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plotTimeWindows}
\alias{plotTimeWindows}
\title{plotTimeWindows}
\usage{
plotTimeWindows(vivid = TRUE, verbose = 1, title = "Movie frameline",
  subtitle = "Time Windows", left = "window width [seconds]")
}
\arguments{
\item{vivid}{frames}

\item{verbose}{frames}

\item{title}{frames}

\item{subtitle}{frames}

\item{left}{frames}
}
\value{
none
}
\description{
plotTimeWindows
}
\examples{
print("example")

}
